Create README.md
A complete and functional Twitter-style application. The Active Record pattern was
used to integrate the database with project classes. Simple templates have been used to
generate messages and tweets. The user has the option of registering for the service, changing
his data and deleting the account. Implemented functionalities: adding tweets visible to all
users, commenting on tweets and sending private messages visible only to the sender and
recipient. Displaying all data of user profile.